check user role

+++ movie_service/app/routes.py
@@ -6,7 +6,7 @@ import requests
 from app.models.movie import Movie
 
 
-AUTHENTICATION_URL = os.environ.get('AUTHENTICATION_URL')
+USER_URL = os.environ.get('USER_URL')
 
 
 @app.route('/')
@@ -20,9 +20,14 @@ def get_movie():
     movies = Movie.objects().all()
     return jsonify(movies), 200
 
+
 @app.route('/movie', methods=['POST'])
 def post_movie():
-    # verify user
+    output = requests.get(f"{USER_URL}user", cookies={'session': request.cookies.get('session')})  # TODO: put it as annotation
+    if output.status_code >= 300:
+        return output.json(), output.status_code
+    if output.json()["role"] != "admin":
+        return "error", 404
     movie = Movie()
     movie.runtime = request.json.get('runtime')
     movie.director = request.json.get('director')
@@ -33,28 +38,34 @@ def post_movie():
     movie.save()
     return movie.to_json(), 200
 
+
 @app.route('/movie/<movie_id>', methods=['GET'])
 def get_movie_from_id(movie_id):
     movie = Movie.objects(id=movie_id).first()
     return movie.to_json(), 200
 
+
 @app.route('/movie/<movie_id>', methods=['PUT'])
 def update_movie(movie_id):
-    output = requests.get(f"{AUTHENTICATION_URL}verify_token", cookies={'session': request.cookies.get('session')})  # TODO: put it as annotation
+    output = requests.get(f"{USER_URL}user", cookies={'session': request.cookies.get('session')})  # TODO: put it as annotation
     if output.status_code >= 300:
         return output.json(), output.status_code
+    if output.json()["role"] != "admin":
+        return "error", 404
     movie = Movie.objects(id=movie_id).first()
     movie.update(**request.json)
     movie.reload()
     return movie.to_json(), 200
 
+
 @app.route('/movie/<movie_id>', methods=['DELETE'])
 def delete_movie(movie_id):
-    output = requests.get(f"{AUTHENTICATION_URL}verify_token", cookies={'session': request.cookies.get('session')})  # TODO: put it as annotation
+    output = requests.get(f"{USER_URL}user", cookies={'session': request.cookies.get('session')})  # TODO: put it as annotation
     if output.status_code >= 300:
         return output.json(), output.status_code
+    if output.json()["role"] != "admin":
+        return "error", 404
     movie = Movie.objects(id=movie_id).first()
-    print(movie)
     movie.delete()
     return '', 204
